requirements.txt
alembic==1.17.1
alembic==1.17.0
Flask==3.1.2
Flask_Login==0.6.3
Flask_Migrate==4.1.0
Flask_Migrate==4.0.5
flask_sqlalchemy==3.1.1
flask_wtf==1.2.2
Flask-Moment==1.2.0
numpy==2.3.5
protobuf==6.33.1
pytest==8.4.2
pytest==8.4.1
python-dotenv==1.2.1
setuptools==80.9.0
SQLAlchemy==2.0.44
Werkzeug==3.1.3
WTForms==3.2.1
